+++ DALayer/Entities/Investigacion.cs
@@ -12,7 +12,7 @@ namespace DALayer.Entities
     {
         public Investigacion() { }
 
-        public Investigacion(string nombre, string descripcion, byte[] foto, int tInicial, int incrementoT)
+        public Investigacion(string nombre, string descripcion, byte[] foto, int tInicial, float incrementoT)
         {
             this.nombre = nombre;
             this.descripcion = descripcion;
