+++ DALayer/Entities/Producto.cs
@@ -16,7 +16,7 @@ namespace DALayer.Entities
         public virtual List<Costo> costos { get; set; }
         public virtual List<Capacidad> capacidad { get; set; }
         public int tiempoInicial { get; set; }
-        public int incrementoTiempo { get; set; }
+        public float incrementoTiempo { get; set; }
         public virtual List<Produce> produce { get; set; }
 
         public void addCosto(Costo c)
