+++ SharedEntities/Entities/Investigacion.cs
@@ -9,7 +9,7 @@ namespace SharedEntities.Entities
     public class Investigacion : Producto
     {
         public Investigacion(int id, string name, string description, byte[] photo, List<Costo> cost, List<Capacidad> capacidad,
-            List<Produce> produce, int tInicial, int incrementoT)
+            List<Produce> produce, int tInicial, float incrementoT)
         {
             this.id = id;
             this.nombre = name;
